Spot-AR-main/SpotControls/Python/LEGACY/requirements.txt
bosdyn-api==3.3.0
bosdyn-choreography-client==3.3.0
bosdyn-choreography-protos==3.3.0
bosdyn-client==3.3.0
bosdyn-core==3.3.0
bosdyn-mission==3.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
Deprecated==1.2.14
glfw==2.5.9
grpcio==1.54.2
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
protobuf==4.23.3
pygame==2.4.0
PyJWT==2.7.0
PyOpenGL==3.1.7
pyopenxr==1.0.2701
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
requests==2.31.0
scipy==1.10.1
urllib3==2.0.3
windows-curses==2.3.1
wrapt==1.15.0
XInput-Python==0.4.0
